refactor: (#18) Chat 관련 타입 shared 레이어로 이동
