+++ src/entities/chat/model/index.ts
@@ -1,3 +1,2 @@
-export * from './types';
 export * from './useGameChat';
 export * from './useMainChat';
