+++ src/shared/api/socket/socketContext.ts
@@ -1,6 +1,6 @@
 import { createContext, useContext } from 'react';
 import type { Socket } from 'socket.io-client';
-import type { ChatMessage } from '@/entities/chat';
+import type { ChatMessage } from '@/shared/model/types';
 
 export interface SocketContextProps {
   socket: Socket | null;
